backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.2
supabase==2.5.0
python-jose==3.3.0
passlib==1.7.4
pytest==8.3.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
respx==0.21.1
slowapi==0.1.9
cryptography==46.0.5
aiocache==0.12.3
sentry-sdk==2.56.0
psutil==7.2.2